Move everything into a single module
