--- packages/miniplex-core/src/types.ts
@@ -1,23 +1,0 @@
-export type Predicate<E, D extends E> =
-  | ((v: E) => v is D)
-  | ((entity: E) => boolean)
-
-/**
- * A utility type that marks the specified properties as required.
- */
-export type With<E, P extends keyof E> = E & Required<Pick<E, P>>
-
-export type Without<E, P extends keyof E> = Omit<E, P>
-
-/**
- * A utility type that removes all optional properties.
- */
-export type Strict<T> = WithoutOptional<T>
-
-/* Utility types */
-
-type OptionalKeys<T> = {
-  [K in keyof T]-?: undefined extends T[K] ? K : never
-}
-
-type WithoutOptional<T> = Pick<T, Exclude<keyof T, OptionalKeys<T>[keyof T]>>
